moved metag from separate module to config and env v2 schemas; refactored references
