--- dynatrace/configuration_v1/metag.py
@@ -1,25 +1,0 @@
-"""
-Copyright 2021 Dynatrace LLC
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-"""
-
-from dynatrace.dynatrace_object import DynatraceObject
-
-
-class METag(DynatraceObject):
-    def _create_from_raw_data(self, raw_element: dict):
-        self.context: str = raw_element.get("context")
-        self.key: str = raw_element.get("key")
-        self.value: str = raw_element.get("value")
-        self.string_representation: str = raw_element.get("stringRepresentation")
