+++ test/environment_v2/test_entities.py
@@ -11,8 +11,7 @@ from dynatrace.environment_v2.monitored_entities import (
     MessageType,
     CustomDeviceCreation,
 )
-from dynatrace.environment_v2.schemas import ManagementZone
-from dynatrace.configuration_v1.metag import METag
+from dynatrace.environment_v2.schemas import ManagementZone, METag
 
 from dynatrace.pagination import PaginatedList
 from dynatrace.utils import int64_to_datetime
